Move the order dependency calculating functions to engine/compare.py
